Improve MySnaps updating
- wait for authorization
- re-use the updates splash screen with the snapcraft icon
- prepare for migration to updates page (snap tab)
